Added the options menu and finished implementing the percent verification
Removed the finish function since I found an easier way to perform the end of the program.
Added the actions menu where the user will chose what action will be performed.
Ended implementing the percent verification since there was no check if the weights added to a hundred as it should.